+++ src/test/java/racingcar/AcceptanceTest.java
@@ -3,7 +3,7 @@ package racingcar;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import racingcar.controller.RacingCarsController;
-import racingcar.converter.RequestConverter;
+import racingcar.controller.converter.RequestConverter;
 import racingcar.service.RacingCarsService;
 import racingcar.service.dto.RaceProcessResponseDto;
 import racingcar.service.dto.RaceWinnerResponseDto;
